Move sign in related code back to `GdgActivity` If the shared prefs says signed in and we are not signed in. Request a sign in again.
